Correct Subscribe promise flow & rename requestPermission to handlePermissionRequest & rename permissions.ts->permission.ts
