+++ src/argon.ts
@@ -42,7 +42,7 @@ export * from './utils'
 export * from './view'
 export * from './visibility'
 export * from './vuforia'
-export * from './permissions'
+export * from './permission'
 export {
     RealityViewer,
     EmptyRealityViewer,
