+++ src/permission.d.ts
@@ -1,0 +1,5 @@
+export declare const PermissionTypes: string[];
+export interface PermissionRequest {
+    type: string;
+    uri: string | undefined;
+}
