--- src/permissions.ts
@@ -1,7 +1,0 @@
-//import { SessionPort } from './session'
-export const PermissionTypes = ['ar.stage', 'ar.camera', 'ar.3dmesh'];
-export interface PermissionRequest {
-    type: string,
-    uri: string | undefined,
-    force?: boolean
-}
